Move llvm-slicer-preprocess.h to tools/include/dg/tools
As the rest of these headers.
